+++ src/user/user-data-initializer.service.ts
@@ -1,19 +1,16 @@
-import {
-  Inject,
-  Injectable,
-  OnModuleInit
-} from '@nestjs/common';
+import { Inject, Injectable, OnModuleInit } from '@nestjs/common';
 import { Model } from 'mongoose';
 import { USER_MODEL } from '../database/database.constants';
 import { RoleType } from '../shared/enum/role-type.enum';
 import { User } from '../database/user.model';
 
 @Injectable()
-export class UserDataInitializerService
-  implements OnModuleInit {
-  constructor(@Inject(USER_MODEL) private userModel: Model<User>) { }
+export class UserDataInitializerService implements OnModuleInit {
+  constructor(@Inject(USER_MODEL) private userModel: Model<User>) {}
 
   async onModuleInit(): Promise<void> {
+    if (process.env.SEED_DATABASE !== 'true') return;
+
     console.log('(UserModule) is initialized...');
     await this.userModel.deleteMany({});
     const user = {
@@ -29,14 +26,9 @@ export class UserDataInitializerService
       email: 'admin@example.com',
       roles: [RoleType.ADMIN],
     };
-    await Promise.all(
-      [
-        this.userModel.create(user),
-        this.userModel.create(admin)
-      ]
-    ).then(
-      data => console.log(data)
-    );
+    await Promise.all([
+      this.userModel.create(user),
+      this.userModel.create(admin),
+    ]).then((data) => console.log(data));
   }
-
 }
